+++ core/templates/dev/head/pages/topic-editor-page/modal-templates/edit-topic-thumbnail-modal.template.html
@@ -16,12 +16,13 @@
         <button class="btn btn-secondary oppia-thumbnail-reset-button" ng-click="reset()">
           <i class="material-icons oppia-vcenter">&#xE14C;</i>
         </button>
-        <img-crop image="uploadedImage" result-image="croppedImageDataUrl"
+        <!-- <img-crop image="uploadedImage" result-image="croppedImageDataUrl"
                   area-type="square" result-image-size="150"
                   on-load-error="onInvalidImageLoaded()"
                   result-image-format="image/png"
                   class="protractor-test-photo-crop">
-        </img-crop>
+        </img-crop> -->
+        <img ng-if="uploadedImage" id="croppable-thumbnail" ng-src="<[uploadedImage]>">
       </div>
     </div>
   </div>
@@ -29,28 +30,28 @@
 
 <div class="modal-footer">
   <button class="btn btn-secondary" ng-click="cancel()" translate="I18N_MODAL_CANCEL_BUTTON"></button>
-  <!--
-      The two checks for ng-disabled are necessary. The former is needed because
-      img-crop loads a default white image even when nothing is uploaded. The latter
-      is needed to prevent the saving of invalid files.
-  -->
   <button class="btn btn-success protractor-test-photo-upload-submit"
           ng-click="confirm()"
-          ng-disabled="!uploadedImage || !croppedImageDataUrl">
+          ng-disabled="!uploadedImage">
     Add Thumbnail
   </button>
 </div>
 <style>
   .oppia-thumbnail-crop-area {
     background: #E4E4E4;
-    height: 350px;
     margin-top: 20px;
+    max-width: 100%;
     position: relative;
-    width: 500px;
+    width: 720px;
   }
   .oppia-thumbnail-reset-button {
     position: absolute;
     right: -50px;
     top: 0;
+  }
+  .oppia-thumbnail-confirm-button {
+    position: absolute;
+    right: -50px;
+    top: 40px;
   }
 </style>
