+++ package.json
@@ -56,6 +56,7 @@
     "cache-loader": "^4.1.0",
     "clean-webpack-plugin": "^2.0.1",
     "core-js": "^2.5.3",
+    "cropperjs": "^1.5.6",
     "css-loader": "^3.1.0",
     "d3": "^5.9.2",
     "dotenv": "^8.1.0",
@@ -75,6 +76,7 @@
     "jasmine-core": "^3.5.0",
     "jasmine-spec-reporter": "^4.2.1",
     "jquery": "^3.4.0",
+    "jquery-cropper": "^1.0.1",
     "karma": "^4.3.0",
     "karma-chrome-launcher": "^2.2.0",
     "karma-coverage": "^2.0.1",
